refactor: token icon api
