+++ packages/web3-providers/src/cf/TokenIconSpecialIconList.json
