+++ packages/web3-providers/src/index.ts
@@ -29,6 +29,7 @@ import { X2Y2API } from './x2y2'
 import { ChainbaseHistoryAPI, ChainbaseFungibleTokenAPI, ChainbaseNonFungibleTokenAPI } from './chainbase'
 import { SolanaFungibleAPI, SolanaNonFungibleAPI } from './solana'
 import { FlowFungibleAPI } from './flow'
+import { CF_API } from './cf'
 
 export * from './helpers'
 export * from './types'
@@ -64,6 +65,7 @@ export const CoinMarketCap = new CoinMarketCapAPI()
 export const UniSwap = new UniSwapAPI()
 export const Rabby = new RabbyAPI()
 export const X2Y2 = new X2Y2API()
+export const CF = new CF_API()
 
 // Alchemy
 export const AlchemyEVM = new AlchemyEVM_API()
