+++ packages/web3-providers/src/types/TokenIcon.ts
@@ -1,0 +1,5 @@
+export namespace TokenIconAPI {
+    export interface Provider<ChainId> {
+        getFungibleTokenIconURLs?: (chainId: ChainId, address: string) => Promise<string[]>
+    }
+}
