+++ packages/web3-providers/src/types/index.ts
@@ -14,5 +14,6 @@ export * from './Security'
 export * from './Storage'
 export * from './Token'
 export * from './TokenList'
+export * from './TokenIcon'
 export * from './Trending'
 export * from './Twitter'
