+++ packages/web3-shared/base/src/utils/resolver.ts
@@ -185,6 +185,7 @@ export const resolveSourceName = createLookupTableResolver<SourceType, string>(
         [SourceType.R2D2]: 'R2D2',
         [SourceType.Rabby]: 'Rabby',
         [SourceType.CoinGecko]: 'CoinGecko',
+        [SourceType.CF]: 'CloudFlare',
     },
     (providerType) => {
         throw new Error(`Unknown provider type: ${providerType}.`)
